requirements.txt
fastapi
uvicorn[standard]
tortoise-orm
asyncpg
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
